fix: 중복 테스트 코드 삭제 및 위치변경
